use risk service in cat review page and refactor
