+++ server/routes/form.js
@@ -99,8 +99,7 @@ module.exports = function Index({
       const form = 'escapeRating'
       const { bookingId } = req.params
       const result = await buildFormData(res, req, section, form, bookingId, transactionalDbClient)
-      const escapeRisk = await riskService.getActiveEscapeRisk(result.data.details.offenderNo, res.locals)
-      const escapeProfile = transformDataToEscapeProfile(escapeRisk)
+      const escapeProfile = await riskService.getEscapeProfile(result.data.details.offenderNo, res.locals)
       const data = { ...result.data, escapeProfile }
       res.render(`formPages/${section}/${form}`, { ...result, data })
     }),
@@ -150,8 +149,7 @@ module.exports = function Index({
       const history = await offendersService.getCatAInformation(res.locals, result.data.details.offenderNo, bookingId)
       const offences = await offendersService.getOffenceHistory(res.locals, result.data.details.offenderNo)
 
-      const escapeProfile = await riskProfilerService.getEscapeProfile(result.data.details.offenderNo, res.locals)
-      console.log(escapeProfile, '<-- escape profile 2')
+      const escapeProfile = await riskService.getEscapeProfile(result.data.details.offenderNo, res.locals)
       const extremismProfile = await riskProfilerService.getExtremismProfile(
         result.data.details.offenderNo,
         res.locals,
