+++ server/utils/utils.js
@@ -1,10 +1,6 @@
 const moment = require('moment')
 const R = require('ramda')
 const { config } = require('../config')
-const {
-  ESCAPE_LIST_ALERT_CODE,
-  ESCAPE_RISK_ALERT_CODE,
-} = require('../data/prisonerSearch/alert/prisonerSearchAlert.dto')
 
 const { dpsUrl, femalePrisonIds } = config
 
@@ -220,30 +216,6 @@ const isOpenCategory = cat => {
 
 const removeLeadingZerosFromDate = date => date.replace(/^0+/, '')?.replace(/\/0/g, '/')
 
-const filterForEscapeAlert = (data, code) => data.filter(item => item.alertCode?.code === code)
-
-const summariseEscapeAlerts = alerts =>
-  alerts.map(item => ({
-    alertCode: item.alertCode?.code,
-    dateCreated: item.activeFrom,
-  }))
-
-const transformDataToEscapeProfile = data => {
-  const activeEscapeListAlerts = filterForEscapeAlert(data, ESCAPE_LIST_ALERT_CODE)
-  const activeEscapeRiskAlerts = filterForEscapeAlert(data, ESCAPE_RISK_ALERT_CODE)
-
-  const activeEscapeListAlertsSummary = summariseEscapeAlerts(activeEscapeListAlerts)
-  const activeEscapeRiskAlertsSummary = summariseEscapeAlerts(activeEscapeRiskAlerts)
-
-  return {
-    activeEscapeList: activeEscapeListAlerts.length > 0,
-    activeEscapeRisk: activeEscapeRiskAlerts.length > 0,
-    esacpeListAlerts: activeEscapeListAlertsSummary,
-    escapeRiskAlerts: activeEscapeRiskAlertsSummary,
-    riskType: 'ESCAPE',
-  }
-}
-
 module.exports = {
   dateConverter,
   dateConverterWithoutLeadingZeros,
@@ -280,5 +252,4 @@ module.exports = {
   // exposed for test purposes
   isBlank,
   removeLeadingZerosFromDate,
-  transformDataToEscapeProfile,
 }
