--- server/utils/utils.test.js
@@ -1,1 +1,0 @@
-// add tests for transformDataToEscapeProfile
